Move the channel maps from src to include
